app/requirements.txt
certifi==2019.11.28
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.9
django-appconf==1.0.3
django-ckeditor==5.8.0
django-debug-toolbar==2.1
django-imagekit==4.0.2
django-import-export==2.0.1
django-js-asset==1.2.2
django-probes==1.2.0
django-taggit==1.2.0
django-treebeard==4.3.1
django-user-agents==0.3.2
djangorestframework==3.11.0
docopt==0.6.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
num2words==0.5.10
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3
sentry-sdk==0.14.0
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
ua-parser==0.8.0
Unidecode==1.1.1
urllib3==1.25.7
user-agents==2.0
xlrd==1.2.0
xlwt==1.3.0
